+++ backend/domain/store/question.go
@@ -13,5 +13,5 @@ type QuestionStore interface {
 	GetQuestionsByQuizID(ctx context.Context, quizID uuid.UUID) ([]*model.Question, error)
 	GetOrderedQuestionsByQuizID(ctx context.Context, quizID uuid.UUID) ([]*model.Question, error)
 	GetQuestionByIDAndQuizID(ctx context.Context, questionID uuid.UUID, quizID uuid.UUID) (*model.Question, error)
-	GetCountQuestionsByQuizID(ctx context.Context, quizID uuid.UUID) (int64, error)
+	GetCountQuestionsByQuizID(ctx context.Context, quizID uuid.UUID) (int, error)
 }
