+++ backend/domain/store/session.go
@@ -8,7 +8,7 @@ import (
 )
 
 type SessionStore interface {
-	CreateSession(ctx context.Context, userID uuid.UUID, quizID uuid.UUID, questionCount int64) (*model.Session, error)
+	CreateSession(ctx context.Context, userID uuid.UUID, quizID uuid.UUID, questionCount int) (*model.Session, error)
 	FindActive(ctx context.Context, userID uuid.UUID, quizID uuid.UUID) (*model.Session, error)
 	FindByID(ctx context.Context, sessionID uuid.UUID) (*model.Session, error)
 	UpdateSession(ctx context.Context, session *model.Session) error
